Use final plugin structure and name schema
- added url for OpenProject project
