+++ lib/open_project/help_link/version.rb
@@ -14,24 +14,8 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-module OpenprojectHelpLink
-  module InfoPatch
-    def self.included(base)
-      base.send(:extend, ClassMethods)
-
-      base.class_eval do
-        unloadable
-
-        class << self
-          alias_method_chain :help_url, :settings
-        end
-      end
-    end
-
-    module ClassMethods
-      def help_url_with_settings
-        Setting.plugin_openproject_help_link["help_link_target"]
-      end
-    end
+module OpenProject
+  module HelpLink
+    VERSION = "3.0.2"
   end
 end
