+++ lib/openproject-help_link.rb
@@ -14,6 +14,4 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-module HelpLink
-  VERSION = "3.0.2"
-end
+require 'open_project/help_link'
